requirements.txt
asgiref==3.10.0
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
drf-spectacular==0.27.2
PyMySQL==1.1.2
python-dotenv==1.2.1
sqlparse==0.5.3
mkdocs==1.5.3
mkdocs-material==9.5.3
mkdocs-minify-plugin==0.8.0
# WebAuthn/Passkey support
webauthn==2.3.0
cbor2==5.6.5
# MCP Server support
mcp==1.1.2
# Testing dependencies
pytest-asyncio==0.23.5
# HTTP transport for MCP server
uvicorn==0.32.1
starlette==0.41.3
# Sentry error tracking
sentry-sdk[django]==2.21.0
